refactor(donate): change amount handling to use decimal values
- Update transaction amount to use decimal type in database
- Modify payment gateway interface to accept float amounts
- Remove number formatting from receipt view
- Adjust donation plan amounts and seeding values

+++ app/Modules/Donate/Domain/Interfaces/PaymentGatewayInterface.php
@@ -8,7 +8,7 @@ interface PaymentGatewayInterface
 {
     public function id(): string;
     public function displayName(): string;
-    public function createCheckout(int $amount, array $meta = []): array;
+    public function createCheckout(float $amount, array $meta = []): array;
     public function handleCallback(Request $request): array;
     public function handleWebhook(Request $request): array;
 }
